General Block Setting for Require Membership  * Add toggle to either show or not content visibility component.  * Add render block filter  * Refactor membership/render.php
