+++ blocks/src/block-membership-required-component/index.js
@@ -1,6 +1,6 @@
 import { __ } from '@wordpress/i18n';
 
-import InspectorControlsFragment from '../membership/inspectorControlsFragment';
+import MembershipContentControls from '../membership/membershipContentControls';
 
 /**
  * Add the visibility attributes to the block settings.
@@ -12,21 +12,25 @@ function addVisibilityAttribute(settings, name) {
 	if (typeof settings.attributes !== 'undefined') {
 		if (name.startsWith('core/')) {
 			settings.attributes = Object.assign(settings.attributes, {
-				invert_restrictions: {
-					type: "string",
-					default: "0"
+				"visibilityBlockEnabled": {
+					"type": 'boolean',
+					"default": false
 				},
-				segment:{
-					type: "string",
-					default: "all"
+				"invert_restrictions": {
+					"type": "string",
+					"default": "0"	
 				},
-				levels: {
-					type: "array",
-					default:[]
+				"segment":{
+					"type": "string",
+					"default": "all"
 				},
-				show_noaccess: {
-					type: "string",
-					default: "0"
+				"levels": {
+					"type": "array",
+					"default":[]
+				},
+				"show_noaccess": {
+					"type": "string",
+					"default": "0"
 				}
 			});
 		}
@@ -55,7 +59,7 @@ const membershipRequiredComponent = wp.compose.createHigherOrderComponent((Block
 		 return (	
 			<Fragment>
 				<BlockEdit {...props} />
-				{ isSelected && (props.name.startsWith('core/')) &&	InspectorControlsFragment(props) }
+				{ isSelected && (props.name.startsWith('core/')) &&	MembershipContentControls(props) }
 			</Fragment>
 		);
